requirements.txt
einops==0.5.0
matplotlib==3.7.2
numpy==1.22.4
pandas==1.4.3
PyYAML==6.0
torch==1.12
torchvision==0.13.1
tqdm==4.64.0
